+++ src/dotnet/aspire-session-host/Program.cs
@@ -34,6 +34,9 @@ await sessionEventService.Subscribe();
 var sessionMetricService = new SessionMetricService(connection);
 await sessionMetricService.Subscribe();
 
+var sessionNodeService = new SessionNodeService(connection);
+await sessionNodeService.Subscribe();
+
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddGrpc();
@@ -41,6 +44,7 @@ if (otlpEndpointUrl != null) builder.Services.AddOtelClients(otlpEndpointUrl);
 builder.Services.AddSingleton(connection);
 builder.Services.AddSingleton(sessionEventService);
 builder.Services.AddSingleton(sessionMetricService);
+builder.Services.AddSingleton(sessionNodeService);
 builder.Services.AddSingleton<SessionService>();
 
 builder.Services.ConfigureHttpJsonOptions(it =>